app top

+++ app/src/main/java/com/ji/tree/app/local/AppData.java
@@ -1,11 +1,10 @@
 package com.ji.tree.app.local;
 
 public class AppData {
-
     public String iconUrl;
     public String name;
     public String packageName;
-    public int versionCode;
+    public long versionCode;
 
     @Override
     public String toString() {
